feat(sessions): Add the ingest incremental sums and session update (#174)

+++ src/backend/Zeeq.Platform.Telemetry/Read/AgentConversationEndpoints.Contracts.cs
@@ -48,8 +48,8 @@ public sealed record AgentConversationListItemDto(
     string? CreatedById,
     DateTimeOffset StartedAtUtc,
     DateTimeOffset? CompletedAtUtc,
-    int TotalInputTokens,
-    int TotalOutputTokens,
+    long TotalInputTokens,
+    long TotalOutputTokens,
     decimal? TotalCostUsd
 );
